add labels to guitar tone mode
