+++ Midibard/Configuration/ConfigurationPrivate.cs
